Added FutureCollection object to executor, updated docstrings and moved Executor class out of init
